ProjectsExplorer: Link action buttons

+++ appinventor/appengine/src/com/google/appinventor/client/views/projects/ProjectsExplorer.java
@@ -10,6 +10,7 @@ import com.google.gwt.uibinder.client.UiField;
 import com.google.gwt.uibinder.client.UiHandler;
 import com.google.gwt.user.client.ui.Composite;
 import com.google.gwt.user.client.ui.FlowPanel;
+import com.google.gwt.user.client.Window;
 
 import com.google.appinventor.client.Ode;
 import com.google.appinventor.client.OdeMessages;
@@ -17,19 +18,31 @@ import com.google.appinventor.client.components.Button;
 import com.google.appinventor.client.components.Dropdown;
 import com.google.appinventor.client.explorer.project.Project;
 import com.google.appinventor.client.utils.Downloader;
+import com.google.appinventor.client.wizards.ProjectUploadWizard;
+import com.google.appinventor.client.wizards.TemplateUploadWizard;
 import com.google.appinventor.client.wizards.youngandroid.NewYoungAndroidProjectWizard;
 import com.google.appinventor.shared.rpc.ServerLayout;
+
+import java.util.List;
 
 public class ProjectsExplorer extends Composite {
   interface ProjectsExplorerUiBinder extends UiBinder<FlowPanel, ProjectsExplorer> {}
   private static final ProjectsExplorerUiBinder UI_BINDER = GWT.create(ProjectsExplorerUiBinder.class);
 
+  @UiField FlowPanel projectsViewActions;
+  @UiField FlowPanel trashViewActions;
+
   @UiField Button newProjectButton;
+  @UiField Button importProjectButton;
+  @UiField Button importFromTemplateButton;
 
   @UiField Button downloadButton;
   @UiField Button exportButton;
   @UiField Button publishButton;
 
+  @UiField Button trashButton;
+
+  @UiField Button restoreButton;
   @UiField Button deleteButton;
 
   @UiField FlowPanel exportDropdownContainer;
@@ -40,9 +53,11 @@ public class ProjectsExplorer extends Composite {
   Dropdown mobileDropdown = new Dropdown(true);
 
   @UiField ProjectsList projectsList;
+  @UiField ProjectsList trashList;
 
   public ProjectsExplorer() {
     initWidget(UI_BINDER.createAndBindUi(this));
+    switchToProjects();
     if(Ode.isMobile()) {
       newProjectButton.setText("");
     }
@@ -51,18 +66,31 @@ public class ProjectsExplorer extends Composite {
       public void onSelectionChange(int selectedItemCount) {
         if(selectedItemCount == 0) {
           downloadButton.setEnabled(false);
-          exportButton.setEnabled(false);
-          deleteButton.setEnabled(false);
+          trashButton.setEnabled(false);
         } else {
           downloadButton.setEnabled(true);
-          exportButton.setEnabled(true);
-          deleteButton.setEnabled(true);
+          trashButton.setEnabled(true);
         }
 
         if(selectedItemCount == 1) {
           publishButton.setEnabled(true);
+          exportButton.setEnabled(true);
         } else {
           publishButton.setEnabled(false);
+          exportButton.setEnabled(false);
+        }
+      }
+    });
+
+    trashList.setSelectionChangeHandler(new ProjectsList.SelectionChangeHandler() {
+      @Override
+      public void onSelectionChange(int selectedItemCount) {
+        if(selectedItemCount == 0) {
+          restoreButton.setEnabled(false);
+          deleteButton.setEnabled(false);
+        } else {
+          restoreButton.setEnabled(true);
+          deleteButton.setEnabled(true);
         }
       }
     });
@@ -71,6 +99,20 @@ public class ProjectsExplorer extends Composite {
     mobileDropdown.setWidget(mobileDropdownContainer);
     exportDropdown.setDropdownButton(exportButton);
     exportDropdown.setWidget(exportDropdownContainer);
+  }
+
+  public void switchToTrash() {
+    projectsViewActions.setVisible(false);
+    projectsList.setVisible(false);
+    trashViewActions.setVisible(true);
+    trashList.setVisible(true);
+  }
+
+  public void switchToProjects() {
+    projectsViewActions.setVisible(true);
+    projectsList.setVisible(true);
+    trashViewActions.setVisible(false);
+    trashList.setVisible(false);
   }
 
   @UiFactory
@@ -85,6 +127,16 @@ public class ProjectsExplorer extends Composite {
       return;
     }
     new NewYoungAndroidProjectWizard(null).center();
+  }
+
+  @UiHandler("importProjectButton")
+  void importProject(ClickEvent e) {
+    new ProjectUploadWizard().center();
+  }
+
+  @UiHandler("importFromTemplateButton")
+  void importFromTemplate(ClickEvent e) {
+    new TemplateUploadWizard().center();
   }
 
   @UiHandler("downloadButton")
@@ -102,12 +154,84 @@ public class ProjectsExplorer extends Composite {
     }
 
     Downloader.getInstance().download(selectedProjPath);
+    projectsList.setSelected(false);
+  }
+
+  @UiHandler("publishButton")
+  void publishProject(ClickEvent e) {
+    Project project = projectsList.getSelectedProjects().get(0);
+    projectsList.setSelected(false);
+    Ode.getInstance().getProjectService().sendToGallery(project.getProjectId(),
+      new OdeAsyncCallback<RpcResult>(
+        MESSAGES.GallerySendingError()) {
+        @Override
+        public void onSuccess(RpcResult result) {
+          if (result.getResult() == RpcResult.SUCCESS) {
+            Window.open(result.getOutput(), "_blank", "");
+          } else {
+            ErrorReporter.reportError(result.getError());
+          }
+        }
+        @Override
+        public void onFailure(Throwable t) {
+          super.onFailure(t);
+        }
+      });
+  }
+
+  @UiHandler("trashButton")
+  void trashSelectedProjects(ClickEvent e) {
+    if (confirmDelete(projectsList.getSelectedProjects())) {
+      for(Project project : projectsList.getSelectedProjects()) {
+        project.moveToTrash();
+      }
+      projectsList.setSelected(false);
+      trashList.setSelected(false);
+    }
+  }
+
+  @UiHandler("restoreButton")
+  void restoreSelectedProjects(ClickEvent e) {
+    for(Project project : trashList.getSelectedProjects()) {
+      project.restoreFromTrash();
+    }
+    projectsList.setSelected(false);
+    trashList.setSelected(false);
   }
 
   @UiHandler("deleteButton")
   void deleteSelectedProjects(ClickEvent e) {
-    for(Project project : projectsList.getSelectedProjects()) {
-      project.moveToTrash();
+    for(Project project : trashList.getSelectedProjects()) {
+      project.deleteFromTrash();
     }
+    projectsList.setSelected(false);
+    trashList.setSelected(false);
+  }
+
+  @UiHandler("viewTrashButton")
+  void viewTrash(ClickEvent e) {
+    switchToTrash();
+  }
+
+  @UiHandler("viewProjectsButton")
+  void viewProjects(ClickEvent e) {
+    switchToProjects();
+  }
+
+  private boolean confirmDelete(List<Project> projects) {
+    String message;
+    if (projects.size() == 1) {
+      message = MESSAGES.confirmMoveToTrashSingleProject(projects.get(0).getProjectName());
+    } else {
+      StringBuilder sb = new StringBuilder();
+      String separator = "";
+      for (Project project : projects) {
+        sb.append(separator).append(project.getProjectName());
+        separator = ", ";
+      }
+      String projectNames = sb.toString();
+      message = MESSAGES.confirmMoveToTrash(projectNames);
+    }
+    return Window.confirm(message);
   }
 }
